requirements.txt
anyio==3.6.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.5
dnspython==2.3.0
facebook-sdk==3.1.0
fastapi==0.81.0
h11==0.13.0
httptools==0.5.0
idna==3.3
npm==0.1.1
optional-django==0.1.0
pydantic==1.9.2
pymongo==4.3.3
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.2
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.14
uvicorn==0.18.3
watchfiles==0.18.1
websockets==10.4
